Tested with defined inputs.
