+++ localization/CMakeLists.txt
@@ -1,0 +1,5 @@
+cmake_minimum_required(VERSION 3.1)
+project(localization)
+
+include_directories(include)
+add_executable(test test.cpp src/localization.cpp)
